feat : variants attributes
