+++ tests/Integration/Attributes/AttributesCest.php
@@ -7,27 +7,13 @@ namespace Kanvas\Inventory\Tests\Integration\Attributes;
 use IntegrationTester;
 use Kanvas\Inventory\Attributes\Attribute;
 use Kanvas\Inventory\Attributes\Models\Attributes;
-use Kanvas\Inventory\Categories\Category;
 use Kanvas\Inventory\Enums\State;
 use Kanvas\Inventory\Tests\Support\Models\Users;
+use Kanvas\Inventory\Tests\Support\Traits\CanCreateAttribute;
 
 class AttributesCest
 {
-    /**
-     * Create a category.
-     *
-     * @return Category
-     */
-    protected function createAttribute(IntegrationTester $I) : Attributes
-    {
-        $user = new Users();
-        $attribute = Attribute::create(
-            $user,
-            $I->faker()->name(),
-        );
-
-        return $attribute;
-    }
+    use CanCreateAttribute;
 
     public function testCreate(IntegrationTester $I) : void
     {
